backend/requirements.txt
# ========================================
# Autonomous City Traffic Intelligence System
# Backend Requirements
# ========================================

# ========================================
# CRITICAL: Reinforcement Learning (Core)
# ========================================
stable-baselines3==2.2.1
gym==0.26.2
torch==2.1.0
tensorboard==2.15.1

# ========================================
# FastAPI & Server
# ========================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ========================================
# WebSocket Support
# ========================================
python-socketio==5.10.0
python-engineio==4.8.0

# ========================================
# Database & ORM
# ========================================
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0

# ========================================
# Data Validation
# ========================================
pydantic==2.5.0
pydantic-settings==2.1.0

# ========================================
# Data Processing
# ========================================
numpy==1.26.2
pandas==2.1.3

# ========================================
# Geospatial & Live Traffic APIs (NEW)
# ========================================
osmnx==1.6.0
networkx==3.2
aiohttp==3.9.1
geopy==2.4.1
requests==2.31.0

# ========================================
# Async Support
# ========================================
aiofiles==23.2.1

# ========================================
# Utilities
# ========================================
python-dotenv==1.0.0
python-dateutil==2.8.2
pyyaml==6.0.1

# ========================================
# Development Tools
# ========================================
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.28.1
black==23.11.0
flake8==6.1.0

